+++ kythe/java/com/google/devtools/kythe/extractors/shared/ExtractorUtils.java
@@ -21,7 +21,6 @@ import static com.google.common.hash.Hashing.sha256;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Function;
-import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Lists;